requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
crc32c==2.3.post0
Django==4.2.7
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.3
pkg_resources==0.0.0
pyproject==1.3.1
python-snappy==0.6.1
sqlparse==0.4.4
typing_extensions==4.8.0
